==> notebooks/02_wego_tsp_segmentation (2) - Copy.ipynb <==
## WeGo Route 50 Data TSP Segmentation

import pandas as pd
import numpy as np
from datetime import datetime

wego = pd.read_csv("../data/wego_contextual_segmentation.csv")

### 1. Assigning TSP Phase to Each Trip

These phases include:

- `Pre-TSP` – before any TSP was active (... - 02.03.2025 12:00 pm)

- `TSP Unconditional 1` – full TSP activation (02.03.2025 12:00 pm - 02.10.2025 12:00 pm)

- `TSP Schedule-Conditional` – TSP activates only if bus is late more than 2 min (02.10.2025 12:00 pm - 04.28.2025 12:00 pm)

- `TSP Off` – TSP deactivated (04.28.2025 12:00 pm - 05.05.2025 12:00 pm)

- `TSP Unconditional 2` – full TSP activation (05.12.2025 12:00 pm - 05.05.2025 12:00 pm)

- `TSP Headway-Conditional` – TSP activates if headway deviation exceeds 120% (05.05.2025 12:00 pm - ...)

wego['DATETIME'] = pd.to_datetime(wego['DATE'].astype(str) + ' ' + wego['SCHEDULED_TIME'].astype(str), errors='coerce')

def assign_tsp_phase(dt):
    if pd.isna(dt):
        return ('Pre-TSP')
    if dt < datetime(2025, 2, 3, 12):
        return ('Pre-TSP')
    elif dt < datetime(2025, 2, 10, 12):
        return ('TSP Unconditional 1')
    elif dt < datetime(2025, 4, 28, 12):
        return ('TSP Schedule-Conditional')
    elif dt < datetime(2025, 5, 5, 12):
        return ('TSP Off')
    elif dt < datetime(2025, 5, 12, 12):
        return ('TSP Unconditional 2')
    else:
        return ('TSP Headway-Conditional')

wego['TSP_PHASE_NAME'] = wego['DATETIME'].apply(assign_tsp_phase)

### 2. Assigning TSP_ACTIVE Flag
This function creates a new column TSP_ACTIVE which indicates whether Transit Signal Priority (TSP) was actively applied to a specific timepoint record ('Y' or 'N').

def assign_tsp_active(row):
    phase = row['TSP_PHASE_NAME']

    timepoint = row['TIME_POINT_ABBR']
    
    tsp_timepoints = ['WHBG', 'CH46', '28&CHARL', 'MCC5_1']
    if timepoint not in tsp_timepoints:
        return 'N'
    
    if phase == 'Pre-TSP' or phase == 'TSP Off':
        return 'N'
    
    elif phase in ['TSP Unconditional 1', 'TSP Unconditional 2']:
        return 'Y'
    
    elif phase == 'TSP Schedule-Conditional':
        delay_minutes = row['ADHERENCE']
        
        if delay_minutes < -2:
            return 'Y'  
        else: 'N'

    elif phase == 'TSP Headway-Conditional':
        hdwy = row['HDWY_DEV']
        
        if hdwy > 1.2:
            return 'Y'  
        else: 'N'
            
    else:
        return 'N'

wego['TSP_ACTIVE'] = wego.apply(assign_tsp_active, axis=1)

wego.head()

### 3. Save the cleaned and enriched dataset

wego.to_csv("../data/wego_tsp_assigned.csv", index=False)

cln_wego = wego[(wego['OVERLOAD_ID'] == 0) &
(wego['TRIP_EDGE'] == 0) &
(wego['STOP_CANCELLED'] == 0) &
(~wego['TIME_POINT_ABBR'].str.contains('HLWD', na=False))]

non_nan_values = cln_wego['SCHEDULED_HDWY'].dropna()

non_null_headway = cln_wego[['DATE', 'DAY_TYPE', 'TIME_OF_DAY', 'SEASON', 'IN_SCHOOL_ZONE_TIME', 'IS_HOLIDAY', 'OPERATOR', 'TIME_POINT_ABBR', 'ROUTE_DIRECTION_NAME', 'SCHEDULED_TIME', 'ACTUAL_ARRIVAL_TIME', 'ADHERENCE', 'DWELL_IN_MINS', 'SCHEDULED_HDWY', 'ACTUAL_HDWY', 'HDWY_DEV', 'DATETIME', 'TSP_PHASE_NAME',	'TSP_ACTIVE']].dropna()
non_null_headway 

non_null_headway['HDWY_DEV_PCT'] = (non_null_headway['HDWY_DEV'] / non_null_headway['SCHEDULED_HDWY']) * 100
non_null_headway.sort_values(['HDWY_DEV'])

non_null_headway[(non_null_headway['ADHERENCE'] < -20)].sort_values('ADHERENCE')

non_null_headway
