==> notebooks/03_wego_analysis.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

#importing csv
wego = pd.read_csv(r"C:\Users\ekims\Documents\NSS\DA14\Projects\wego-partnerproject-da14-v2-ekimshaf\data\wego_tsp_assigned.csv")

wego.info()

wego.head()

1) This section begins parsing out the time and date points of TSP activation in the data into 
individual dataframes.

#Sectioning out data from before TSP activation in dataset for route in question.
# TSP began on 2/3/2025 at 12:00pm (noon)
before_tsp = wego[(wego['CALENDAR_ID'] <= 120250203) & 
(wego['TSP_PHASE_NAME'].str.contains('Pre-TSP'))]
before_tsp.info()

before_tsp

##filtering dates that are included in unconditional TSP, Feb 3rd - Feb 10th, 2025 11:59am
tsp_uncon_feb = wego[(wego['CALENDAR_ID'] >= 120250203) & 
(wego['CALENDAR_ID'] <= 120250210) &
(wego['TSP_PHASE_NAME'].str.contains('TSP Unconditional 1'))]
tsp_uncon_feb.info()

tsp_uncon_feb

##filtering dates that are included in conditional TSP, Feb 10th - Apr 28, 2025
tsp_cond_feb = wego[(wego['CALENDAR_ID'] >= 120250210) & 
(wego['CALENDAR_ID'] <= 120250428) &
(wego['TSP_PHASE_NAME'].str.contains('TSP Schedule-Conditional'))]
tsp_cond_feb.info()

tsp_cond_feb

#filtering dates with TSP off between Apr 28th and May 5th
tsp_off_apr = wego[(wego['CALENDAR_ID'] >= 120250428) & 
(wego['CALENDAR_ID'] <= 120250505) &
(wego['TSP_PHASE_NAME'].str.contains('TSP Off'))]
tsp_off_apr.info()

tsp_off_apr

#Filtering dates for headway priority condition for TSP beginning May 12th
tsp_cond_may = wego[(wego['CALENDAR_ID'] >= 120250512) &
(wego['TSP_PHASE_NAME'].str.contains('TSP Headway-Conditional'))]
tsp_cond_may.info()

tsp_cond_may

2. Data has been sectioned out into dataframes for each timepoint:

Dataframe names:

before_tsp  
tsp_uncon_feb  
tsp_cond_feb  
tsp_off_apr  
tsp_cond_may
